// panels/punjabship-panels/courier-cart-client/src/types/generic.types.ts
export type AddressType = 'pickup' | 'rto' | 'billing'

export interface IAddress {
  id?: string
  userId?: string
  type?: AddressType

  contactName?: string
  contactPhone?: string
  contactEmail?: string
  addressNickname?: string
  contactPersonRole?: string

  addressLine1?: string
  addressLine2?: string
  landmark?: string

  city?: string
  state?: string
  country?: string
  pincode?: string

  latitude?: string
  longitude?: string
  gstNumber?: string

  createdAt?: string
  updatedAt?: string
}

export interface HydratedPickup {
  id: string
  userId: string
  addressId: string
  rtoAddressId?: string | null
  pickupId: string
  isPrimary: boolean
  isPickupEnabled: boolean
  pickup: IAddress
  rto?: IAddress | null
  createdAt: string
  isRTOSame: boolean
  updatedAt: string
}